refactor: relocate package under src layout
