+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
 addopts = --strict-markers
+pythonpath = src
 testpaths = tests
 markers =
     integration: exercises filesystem or subprocess behavior
