+++ src/yupp/lib/corolib.yu
