+++ tests/conftest.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from pp import yugen, yup
+from yupp.pp import yugen, yup
 
 
 REPO_ROOT = Path(__file__).resolve().parents[1]
